Commit 1 works around the slow dynamic dispatch that currently exist in Base (https://github.com/JuliaLang/julia/issues/29887). This significantly reduces the overhead of creating the Parser:

Before:

```jl
julia> @btime JSON.parse("\"foo\"")
  3.080 μs (2 allocations: 64 bytes)
"foo"
```

After:
```jl
julia> @btime JSON.parse("\"foo\"")
  196.499 ns (3 allocations: 160 bytes)
"foo"
```

Note that quite a lot of JSON can be parsed within 3μs.

Commit 2, caches the UInt8 vector which is filled with bytes of digits. Reusing the same vector has a bunch of benefits in terms of not having to grow it over and over and the data will be hot.

Using the benchmark

```jl
julia> using Random

julia> a = string("[", join([rand(Int) for i in 1:10^3], ","), "]");
```

Before:

```jl
julia> @btime JSON.parse($a)
  352.473 μs (6011 allocations: 250.86 KiB)
```

After:

```jl
julia> @btime JSON.parse($a)
  228.998 μs (1016 allocations: 32.33 KiB)
```

The third commit doesn't at the moment not have much of a significant performance change but it should be strictly less work for not really more complicated code so why not? :)